Move Cosmos validator sorting logic to Cosmos validator list component
