+++ packages/client/src/tools/client-utils.ts
@@ -15,10 +15,10 @@ import {
 import { ApolloError } from "apollo-client";
 import BigNumber from "bignumber.js";
 import { PORTFOLIO_CHART_TYPES } from "i18n/english";
-import { VALIDATORS_LIST_SORT_FILTER } from "modules/app/store";
 import queryString, { ParsedQuery } from "query-string";
 import { AvailableReward } from "ui/CreateTransactionForm";
 import Toast from "ui/Toast";
+import { VALIDATORS_LIST_SORT_FILTER } from "ui/validators/CosmosValidators";
 import {
   convertCryptoToFiat,
   denomToUnit,
